+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
+# [5.1.0] - 2023-05-07
 # Added
 - Added Localization for Russian, German, French, Chinese, Spanish (machine translated).
 - Added `Underpowered Ability` selection to OP ability selection.
 - Added `Excalibur` Special Power.
 - Added `Unreactable` Special Power.
+- Added `Haggler` Special Power.
 # Changed
 - Updated description for `Gifted` feature.
 - Updated description for `Extra Strike` feature.